refactor(agents): fold lifecycle factories into SubagentLifecycleController
