+++ src/agents/subagents/registry/subagent-registry-lifecycle-cleanup.ts
@@ -1,633 +1,689 @@
+import type { cleanupBrowserSessionsForLifecycleEnd } from "../../../browser-lifecycle-cleanup.js";
+import { runWithoutOwnedSessionTranscriptWrites } from "../../../config/sessions/transcript-write-context.js";
+import {
+  isGatewayRestartDraining,
+  runWithGatewayIndependentRootWorkAdmission,
+} from "../../../process/gateway-work-admission.js";
 import { defaultRuntime } from "../../../runtime.js";
-import { normalizeDeliveryContext } from "../../../utils/delivery-context.shared.js";
+import { emitSessionLifecycleEvent } from "../../../sessions/session-lifecycle-events.js";
+import { recordSubagentTerminalState } from "../../../sessions/session-state-events.js";
+import { retireSessionMcpRuntimeForSessionKey } from "../../agent-bundle-mcp-tools.js";
+import { releaseSwarmRun } from "../swarm/swarm-scheduler.js";
 import {
   ensureCompletionState,
   ensureDeliveryState,
   getDeliveryLastError,
-  isDeliverySuspended,
 } from "./subagent-delivery-state.js";
-import { SUBAGENT_ENDED_REASON_COMPLETE } from "./subagent-lifecycle-events.js";
+import {
+  SUBAGENT_ENDED_REASON_KILLED,
+  type SubagentLifecycleEndedReason,
+} from "./subagent-lifecycle-events.js";
 import { shouldSuppressSubagentRecoverySessionEffects } from "./subagent-recovery-state.js";
 import {
-  resolveCleanupCompletionReason,
-  resolveDeferredCleanupDecision,
-} from "./subagent-registry-cleanup.js";
-import {
-  ANNOUNCE_COMPLETION_HARD_EXPIRY_MS,
-  ANNOUNCE_EXPIRY_MS,
   logAnnounceGiveUp,
   MIN_ANNOUNCE_RETRY_DELAY_MS,
+  persistSubagentSessionTiming,
   resolveAnnounceRetryDelayMs,
-  safeRemoveAttachmentsDir,
 } from "./subagent-registry-helpers.js";
-import { createSubagentRegistryLifecycleBookkeeping } from "./subagent-registry-lifecycle-bookkeeping.js";
-import type { createSubagentRegistryLifecycleCleanupBase } from "./subagent-registry-lifecycle-cleanup-base.js";
-import type { createSubagentRegistryLifecycleCommon } from "./subagent-registry-lifecycle-common.js";
+import {
+  markPendingFinalDelivery,
+  safeMarkRequiredCompletionDeliveryBlocked,
+  safeSetSubagentTaskDeliveryStatus,
+} from "./subagent-registry-lifecycle-delivery.js";
+import {
+  markRequesterSettleWakePending,
+  scheduleRequesterSettleWake,
+} from "./subagent-registry-lifecycle-wake.js";
 import type {
-  SubagentRegistryLifecycleParams,
-  SubagentRegistryLifecycleState,
-} from "./subagent-registry-lifecycle-contracts.js";
-import type { createSubagentRegistryLifecycleDelivery } from "./subagent-registry-lifecycle-delivery.js";
-import type { createSubagentRegistryLifecycleRequesterWake } from "./subagent-registry-lifecycle-requester-wake.js";
-import { loadSubagentSessionEntry } from "./subagent-session-reconciliation.js";
+  SubagentLifecycleCommonContext,
+  SubagentLifecycleCompletionContext,
+  SubagentLifecycleCleanupContext,
+  SubagentLifecycleOptions,
+  SubagentLifecycleWakeContext,
+} from "./subagent-registry-lifecycle.js";
+import type { SubagentCompletionRequest, SubagentRunRecord } from "./subagent-registry.types.js";
 
-type RunSubagentAnnounceFlow =
-  (typeof import("../announce/subagent-announce.js"))["runSubagentAnnounceFlow"];
-type SubagentAnnounceFlowOutcome = Awaited<ReturnType<RunSubagentAnnounceFlow>>;
-import type { SubagentRunRecord } from "./subagent-registry.types.js";
-import { deleteSubagentSessionForCleanup } from "./subagent-session-cleanup.js";
+const MAX_DETACHED_CLEANUP_RETRIES = 3;
 
-export function createSubagentRegistryLifecycleCleanup(
-  params: SubagentRegistryLifecycleParams,
-  state: SubagentRegistryLifecycleState,
-  common: ReturnType<typeof createSubagentRegistryLifecycleCommon>,
-  deliveryHelpers: ReturnType<typeof createSubagentRegistryLifecycleDelivery>,
-  requesterWake: ReturnType<typeof createSubagentRegistryLifecycleRequesterWake>,
-  cleanupBase: ReturnType<typeof createSubagentRegistryLifecycleCleanupBase>,
-) {
-  const { cleanupGenerations } = state;
-  const { buildSafeLifecycleErrorMeta, maskRunId, maskSessionKey } = common;
-  const {
-    clearPendingFinalDelivery,
-    emitCompletionEndedHookIfNeeded,
-    formatAnnounceDeliveryError,
-    hasPriorRequesterDeliveryMirror,
-    loadPendingFinalDeliveryPayload,
-    markPendingFinalDelivery,
-    recordAnnounceDeliveryResult,
-    safeMarkRequiredCompletionDeliveryBlocked,
-    safeSetSubagentTaskDeliveryStatus,
-  } = deliveryHelpers;
-  const {
-    beginSubagentCleanup,
-    isCleanupAttemptCurrent,
-    isEndedHookOwnerCurrent,
-    retireRunModeBundleMcpRuntime,
-    retireSupersededCleanupIfNeeded,
-    retireSupersededCleanupInBackground,
-    runDetachedCleanupAttempt,
-    scheduleResumeSubagentRun,
-    suspendPendingFinalDelivery,
-  } = cleanupBase;
+function isCleanupGenerationCurrent(
+  context: SubagentLifecycleCleanupContext,
+  params: SubagentLifecycleOptions,
+  runId: string,
+  entry: SubagentRunRecord,
+  generation: number,
+): boolean {
+  return (
+    params.runs.get(runId) === entry &&
+    entry.pauseReason !== "sessions_yield" &&
+    context.isCleanupGeneration(entry, generation) &&
+    !context.newerGenerationOwnsSession(entry)
+  );
+}
 
-  const shouldSuspendPendingFinalDelivery = (entry: SubagentRunRecord) =>
-    entry.expectsCompletionMessage === true &&
-    entry.endedReason === SUBAGENT_ENDED_REASON_COMPLETE &&
-    entry.execution.outcome?.status === "ok";
-
-  const finalizeAnnounceGiveUp = async (giveUpParams: {
-    runId: string;
-    entry: SubagentRunRecord;
-    reason: "expiry" | "permanent_failure";
-    cleanup?: "delete" | "keep";
-    cleanupGeneration?: number;
-    retryCount?: number;
-    completedAt?: number;
-  }) => {
-    const { runId, entry, reason, cleanup, cleanupGeneration, retryCount, completedAt } =
-      giveUpParams;
-    if (shouldSuspendPendingFinalDelivery(entry)) {
-      suspendPendingFinalDelivery({
-        runId,
-        entry,
-        reason,
-        error: getDeliveryLastError(entry),
-      });
-      return;
-    }
-    const deliveryError = getDeliveryLastError(entry) ?? reason;
-    clearPendingFinalDelivery(entry);
-    const failedDelivery = ensureDeliveryState(entry);
-    failedDelivery.status = "failed";
-    failedDelivery.lastError = deliveryError;
-    if (retryCount != null) {
-      failedDelivery.attemptCount = retryCount;
-      failedDelivery.lastAttemptAt = completedAt ?? Date.now();
-    }
-    safeSetSubagentTaskDeliveryStatus({
-      entry,
-      deliveryStatus: "failed",
-      deliveryError,
-    });
-    safeMarkRequiredCompletionDeliveryBlocked({
-      entry,
-      reason: deliveryError,
-    });
-    entry.wakeOnDescendantSettle = undefined;
-    const completion = ensureCompletionState(entry);
-    completion.fallbackResultText = undefined;
-    completion.fallbackCapturedAt = undefined;
-    if ((cleanup ?? entry.cleanup) === "delete" || !entry.retainAttachmentsOnKeep) {
-      await safeRemoveAttachmentsDir(entry);
-    }
-    if (
-      cleanupGeneration !== undefined &&
-      !isCleanupAttemptCurrent(runId, entry, cleanupGeneration)
-    ) {
-      await retireSupersededCleanupIfNeeded(runId, entry, cleanupGeneration);
-      return;
-    }
-    const completionReason = resolveCleanupCompletionReason(entry);
-    logAnnounceGiveUp(entry, reason);
-    // Retry-limit / expiry give-up should not leave cleanup stuck behind the
-    // best-effort ended hook. Mark the run cleaned first, then fire the hook.
-    completeCleanupBookkeeping({
-      runId,
-      entry,
-      cleanup: cleanup ?? entry.cleanup,
-      completedAt: completedAt ?? Date.now(),
-    });
-    if (!shouldSuppressSubagentRecoverySessionEffects(entry)) {
-      await emitCompletionEndedHookIfNeeded(
-        entry,
-        completionReason,
-        () =>
-          isEndedHookOwnerCurrent(runId, entry) &&
-          !shouldSuppressSubagentRecoverySessionEffects(entry),
-      );
-    }
-  };
-
-  const retryDeferredCompletedAnnounces = (excludeRunId?: string) => {
-    const now = Date.now();
-    for (const [runId, entry] of params.runs.entries()) {
-      if (excludeRunId && runId === excludeRunId) {
-        continue;
+export function scheduleResumeSubagentRun(
+  context: SubagentLifecycleCleanupContext,
+  params: SubagentLifecycleOptions,
+  runId: string,
+  entry: SubagentRunRecord,
+  delayMs: number,
+  cleanupGeneration?: number,
+): void {
+  const timer = setTimeout(() => {
+    context.deleteScheduledResumeTimer(timer);
+    void runWithGatewayIndependentRootWorkAdmission(async () => {
+      if (params.runs.get(runId) !== entry) {
+        return;
       }
-      if (typeof entry.execution.endedAt !== "number") {
-        continue;
-      }
-      if (entry.cleanupCompletedAt || entry.cleanupHandled) {
-        continue;
-      }
-      if (isDeliverySuspended(entry)) {
-        continue;
-      }
-      if (params.suppressAnnounceForSteerRestart(entry)) {
-        continue;
-      }
-      const endedAgo = now - (entry.execution.endedAt ?? now);
-      if (entry.expectsCompletionMessage !== true && endedAgo > ANNOUNCE_EXPIRY_MS) {
-        if (!beginSubagentCleanup(runId)) {
-          continue;
+      if (cleanupGeneration !== undefined) {
+        if (!isCleanupGenerationCurrent(context, params, runId, entry, cleanupGeneration)) {
+          return;
         }
-        runDetachedCleanupAttempt({
-          runId,
-          entry,
-          cleanupGeneration: cleanupGenerations.get(entry)!,
-          run: async () => {
-            await finalizeAnnounceGiveUp({
-              runId,
-              entry,
-              reason: "expiry",
-            });
-          },
-        });
-        continue;
+        if (entry.cleanupHandled) {
+          entry.cleanupHandled = false;
+          params.persist(runId);
+        }
       }
       params.resumedRuns.delete(runId);
       params.resumeSubagentRun(runId);
-    }
-  };
+    }).catch((err: unknown) => {
+      defaultRuntime.log(`[warn] subagent cleanup resume failed (${runId}): ${String(err)}`);
+      const current = params.runs.get(runId);
+      if (
+        isGatewayRestartDraining() &&
+        current === entry &&
+        typeof current.cleanupCompletedAt !== "number"
+      ) {
+        scheduleResumeSubagentRun(
+          context,
+          params,
+          runId,
+          entry,
+          Math.max(delayMs, MIN_ANNOUNCE_RETRY_DELAY_MS),
+          cleanupGeneration,
+        );
+      }
+    });
+  }, delayMs);
+  timer.unref?.();
+  context.addScheduledResumeTimer(timer);
+}
 
-  const { completeCleanupBookkeeping } = createSubagentRegistryLifecycleBookkeeping(
-    params,
-    common,
-    requesterWake,
-    retryDeferredCompletedAnnounces,
+export function isCleanupAttemptCurrent(
+  context: SubagentLifecycleCleanupContext,
+  params: SubagentLifecycleOptions,
+  runId: string,
+  entry: SubagentRunRecord,
+  generation: number,
+): boolean {
+  return (
+    entry.cleanupHandled === true &&
+    isCleanupGenerationCurrent(context, params, runId, entry, generation)
   );
+}
 
-  const finalizeSubagentCleanup = async (
-    runId: string,
-    cleanup: "delete" | "keep",
-    announceOutcome: SubagentAnnounceFlowOutcome,
-    cleanupGeneration: number,
-    options?: {
-      skipAnnounce?: boolean;
-      skipDeliveryStatus?: boolean;
-      skipRequesterDelivery?: boolean;
-    },
-  ) => {
-    const entry = params.runs.get(runId);
-    if (!entry) {
-      return;
-    }
-    if (!isCleanupAttemptCurrent(runId, entry, cleanupGeneration)) {
-      await retireSupersededCleanupIfNeeded(runId, entry, cleanupGeneration);
-      return;
-    }
-    if (entry.expectsCompletionMessage === false || options?.skipRequesterDelivery) {
-      clearPendingFinalDelivery(entry);
-      if (options?.skipRequesterDelivery) {
-        ensureDeliveryState(entry).status = "not_required";
-        entry.suppressCompletionDelivery = undefined;
-      }
-      entry.wakeOnDescendantSettle = undefined;
-      const shouldDeleteAttachments = cleanup === "delete" || !entry.retainAttachmentsOnKeep;
-      if (shouldDeleteAttachments) {
-        await safeRemoveAttachmentsDir(entry);
-      }
-      if (!isCleanupAttemptCurrent(runId, entry, cleanupGeneration)) {
-        await retireSupersededCleanupIfNeeded(runId, entry, cleanupGeneration);
-        return;
-      }
-      completeCleanupBookkeeping({
-        runId,
-        entry,
-        cleanup,
-        completedAt: Date.now(),
-      });
-      if (!shouldSuppressSubagentRecoverySessionEffects(entry)) {
-        await emitCompletionEndedHookIfNeeded(
-          entry,
-          resolveCleanupCompletionReason(entry),
-          () =>
-            isEndedHookOwnerCurrent(runId, entry) &&
-            !shouldSuppressSubagentRecoverySessionEffects(entry),
+export function runDetachedCleanupAttempt(
+  context: SubagentLifecycleCleanupContext,
+  params: SubagentLifecycleOptions,
+  args: {
+    runId: string;
+    entry: SubagentRunRecord;
+    cleanupGeneration: number;
+    run: () => Promise<void>;
+  },
+): void {
+  // Completion makes the task projection non-blocking before delivery and
+  // cleanup finish. This independent lease bridges that handoff and owns the
+  // full detached attempt, including its final durable registry write.
+  // Completion outlives the spawning attempt; inherited lock owners would
+  // reject requester transcript writes after that attempt is disposed.
+  runWithoutOwnedSessionTranscriptWrites(() => {
+    void runWithGatewayIndependentRootWorkAdmission(async () => {
+      try {
+        await args.run();
+        context.clearCleanupFailureCount(args.entry);
+      } catch (err) {
+        defaultRuntime.log(
+          `[warn] subagent cleanup finalize failed (${args.runId}): ${String(err)}`,
         );
-      }
-      return;
-    }
-    if (announceOutcome === "delivered" || announceOutcome === "intentional_non_delivery") {
-      const delivery = ensureDeliveryState(entry);
-      const shouldCreditDelivery =
-        announceOutcome === "delivered" &&
-        (!options?.skipAnnounce ||
-          delivery.status === "delivered" ||
-          typeof delivery.announcedAt === "number");
-      if (shouldCreditDelivery) {
-        const deliveredAt = delivery.deliveredAt ?? delivery.announcedAt ?? Date.now();
-        delivery.status = "delivered";
-        delivery.deliveredAt = deliveredAt;
-        delivery.announcedAt = delivery.announcedAt ?? deliveredAt;
-        if (!options?.skipAnnounce) {
-          delivery.announcedAt = deliveredAt;
-          params.persist(runId);
-        }
-      }
-      if (announceOutcome === "delivered") {
-        clearPendingFinalDelivery(entry);
-      } else {
-        // The requester-settle batch owns the real delivery now. Retire the
-        // per-child retry obligation without converting the handoff into success.
-        delivery.status = "pending";
-        delivery.disposition = "intentional_non_delivery";
-        delivery.payload = undefined;
-        delivery.createdAt = undefined;
-        delivery.attemptCount = undefined;
-        delivery.nextAttemptAt = undefined;
-      }
-      const finalDelivery = ensureDeliveryState(entry);
-      if (shouldCreditDelivery) {
-        finalDelivery.status = "delivered";
-        finalDelivery.suspendedAt = undefined;
-        finalDelivery.suspendedReason = undefined;
-      }
-      if (shouldCreditDelivery && !options?.skipDeliveryStatus) {
-        safeSetSubagentTaskDeliveryStatus({
-          entry,
-          deliveryStatus: "delivered",
-        });
-      } else if (announceOutcome === "intentional_non_delivery" && !options?.skipDeliveryStatus) {
-        safeSetSubagentTaskDeliveryStatus({ entry, deliveryStatus: "pending" });
-      }
-      if (announceOutcome === "delivered") {
-        finalDelivery.lastError = undefined;
-        finalDelivery.lastDropReason = undefined;
-      }
-      entry.wakeOnDescendantSettle = undefined;
-      const completion = ensureCompletionState(entry);
-      completion.fallbackResultText = undefined;
-      completion.fallbackCapturedAt = undefined;
-      const completionReason = resolveCleanupCompletionReason(entry);
-      const shouldDeleteAttachments = cleanup === "delete" || !entry.retainAttachmentsOnKeep;
-      if (shouldDeleteAttachments) {
-        await safeRemoveAttachmentsDir(entry);
-      }
-      if (!isCleanupAttemptCurrent(runId, entry, cleanupGeneration)) {
-        await retireSupersededCleanupIfNeeded(runId, entry, cleanupGeneration);
-        return;
-      }
-      completeCleanupBookkeeping({
-        runId,
-        entry,
-        cleanup,
-        completedAt: Date.now(),
-      });
-      // Hook loading is best-effort; durable delivery and cleanup must already
-      // be terminal before plugin code can fail or stall.
-      if (!shouldSuppressSubagentRecoverySessionEffects(entry)) {
-        await emitCompletionEndedHookIfNeeded(
-          entry,
-          completionReason,
-          () =>
-            isEndedHookOwnerCurrent(runId, entry) &&
-            !shouldSuppressSubagentRecoverySessionEffects(entry),
-        );
-      }
-      return;
-    }
-
-    if (announceOutcome === "session_queued") {
-      // The correlated queue owns transport now. Settlement, not admission,
-      // decides delivered versus blocked and re-enters cleanup afterward.
-      entry.cleanupHandled = false;
-      params.resumedRuns.delete(runId);
-      params.persist(runId);
-      return;
-    }
-
-    const now = Date.now();
-    const deferredDecision = resolveDeferredCleanupDecision({
-      entry,
-      now,
-      activeDescendantRuns: Math.max(0, params.countPendingDescendantRuns(entry.childSessionKey)),
-      announceExpiryMs: ANNOUNCE_EXPIRY_MS,
-      announceCompletionHardExpiryMs: ANNOUNCE_COMPLETION_HARD_EXPIRY_MS,
-      deferDescendantDelayMs: MIN_ANNOUNCE_RETRY_DELAY_MS,
-      resolveAnnounceRetryDelayMs,
-    });
-
-    if (deferredDecision.kind === "defer-descendants") {
-      ensureDeliveryState(entry).lastAttemptAt = now;
-      entry.wakeOnDescendantSettle = true;
-      entry.cleanupHandled = false;
-      params.resumedRuns.delete(runId);
-      params.persist(runId);
-      scheduleResumeSubagentRun(runId, entry, deferredDecision.delayMs);
-      return;
-    }
-
-    if (deferredDecision.kind === "give-up") {
-      await finalizeAnnounceGiveUp({
-        runId,
-        entry,
-        reason: deferredDecision.reason,
-        cleanup,
-        cleanupGeneration,
-        retryCount: deferredDecision.retryCount,
-        completedAt: now,
-      });
-      return;
-    }
-
-    markPendingFinalDelivery({
-      entry,
-      error: "announce deferred or direct delivery failed",
-    });
-    const delivery = ensureDeliveryState(entry);
-    delivery.windowStartedAt ??= entry.execution.endedAt ?? now;
-    delivery.deadlineAt ??= delivery.windowStartedAt + ANNOUNCE_COMPLETION_HARD_EXPIRY_MS;
-    delivery.nextAttemptAt = now + (deferredDecision.resumeDelayMs ?? 0);
-    entry.cleanupHandled = false;
-    params.resumedRuns.delete(runId);
-    params.persist(runId);
-    if (deferredDecision.resumeDelayMs == null) {
-      return;
-    }
-    scheduleResumeSubagentRun(runId, entry, deferredDecision.resumeDelayMs);
-  };
-
-  const startSubagentAnnounceCleanupFlow = (runId: string, entry: SubagentRunRecord): boolean => {
-    if (entry.killReconciliation) {
-      // Restores and unrelated cleanup retries must not publish a provisional
-      // kill. The sweeper re-enters here after durable reconciliation.
-      return false;
-    }
-    const cleanup = entry.cleanup;
-    let suppressSessionEffects = shouldSuppressSubagentRecoverySessionEffects(entry);
-    if (typeof entry.delivery?.announcedAt === "number" || entry.delivery?.status === "delivered") {
-      if (!beginSubagentCleanup(runId)) {
-        return false;
-      }
-      const cleanupGeneration = cleanupGenerations.get(entry)!;
-      runDetachedCleanupAttempt({
-        runId,
-        entry,
-        cleanupGeneration,
-        run: async () => {
-          await finalizeSubagentCleanup(runId, cleanup, "delivered", cleanupGeneration, {
-            skipAnnounce: true,
-          });
-        },
-      });
-      return true;
-    }
-    if (!beginSubagentCleanup(runId)) {
-      return false;
-    }
-    const cleanupGeneration = cleanupGenerations.get(entry)!;
-    const cleanupSessionEntry = suppressSessionEffects
-      ? undefined
-      : loadSubagentSessionEntry({ childSessionKey: entry.childSessionKey });
-    const cleanupSessionIdentity =
-      cleanupSessionEntry?.sessionId && cleanupSessionEntry.lifecycleRevision
-        ? {
-            sessionId: cleanupSessionEntry.sessionId,
-            lifecycleRevision: cleanupSessionEntry.lifecycleRevision,
-          }
-        : undefined;
-    const suppressChildSessionEffects = () => {
-      suppressSessionEffects = true;
-      if (entry.execution.suppressSessionEffects !== true) {
-        const previousExecution = entry.execution;
-        entry.execution = {
-          ...entry.execution,
-          suppressSessionEffects: true,
-        };
-        try {
-          params.persistOrThrow(runId);
-        } catch (error) {
-          entry.execution = previousExecution;
-          suppressSessionEffects = false;
-          throw error;
-        }
-      }
-    };
-    const childSessionEffectsAllowed = () => {
-      if (!suppressSessionEffects && shouldSuppressSubagentRecoverySessionEffects(entry)) {
-        suppressChildSessionEffects();
-      }
-      return !suppressSessionEffects && isCleanupAttemptCurrent(runId, entry, cleanupGeneration);
-    };
-    const skipRequesterDelivery = entry.suppressCompletionDelivery === true;
-    if (entry.expectsCompletionMessage === false || skipRequesterDelivery) {
-      runDetachedCleanupAttempt({
-        runId,
-        entry,
-        cleanupGeneration,
-        run: async () => {
-          // This driver is detached. Yield once so synchronous successor
-          // registration can invalidate it before sessions.delete is submitted.
-          await Promise.resolve();
-          if (!isCleanupAttemptCurrent(runId, entry, cleanupGeneration)) {
-            await retireSupersededCleanupIfNeeded(runId, entry, cleanupGeneration);
-            return;
-          }
-          if (cleanup === "delete" && childSessionEffectsAllowed()) {
-            if (!cleanupSessionIdentity) {
-              // Without both lifecycle identities, key-only deletion could remove
-              // a successor that reused this child session after cleanup yielded.
-              suppressChildSessionEffects();
-            } else {
-              // This durable boundary prevents a late yield from reviving a run
-              // after deletion may already have reached the gateway.
-              entry.deleteCleanupDispatchedAt ??= Date.now();
-              params.persist(runId);
-              const sessionCleanup = await deleteSubagentSessionForCleanup({
-                callGateway: params.callGateway,
-                childSessionKey: entry.childSessionKey,
-                spawnMode: entry.spawnMode,
-                expectedSessionId: cleanupSessionIdentity.sessionId,
-                expectedLifecycleRevision: cleanupSessionIdentity.lifecycleRevision,
-                onError: (error) =>
-                  params.warn("sessions.delete failed during subagent cleanup", {
-                    error: buildSafeLifecycleErrorMeta(error),
-                    runId: maskRunId(runId),
-                    childSessionKey: maskSessionKey(entry.childSessionKey),
-                  }),
-              });
-              if (sessionCleanup === "failed") {
-                throw new Error("subagent session cleanup did not complete");
-              }
-              if (sessionCleanup === "changed") {
-                suppressChildSessionEffects();
-              }
-            }
-          }
-          if (!isCleanupAttemptCurrent(runId, entry, cleanupGeneration)) {
-            await retireSupersededCleanupIfNeeded(runId, entry, cleanupGeneration);
-            return;
-          }
-          await finalizeSubagentCleanup(runId, cleanup, "delivered", cleanupGeneration, {
-            skipAnnounce: true,
-            skipDeliveryStatus: true,
-            skipRequesterDelivery,
-          });
-        },
-      });
-      return true;
-    }
-    const pendingPayload = loadPendingFinalDeliveryPayload(entry);
-    const requesterOrigin = normalizeDeliveryContext(pendingPayload.requesterOrigin);
-    let latestDeliveryError = getDeliveryLastError(entry);
-    const finalizeAnnounceCleanup = async (announceOutcome: SubagentAnnounceFlowOutcome) => {
-      if (!isCleanupAttemptCurrent(runId, entry, cleanupGeneration)) {
-        await retireSupersededCleanupIfNeeded(runId, entry, cleanupGeneration);
-        return;
-      }
-      const shouldCreditPriorDelivery =
-        announceOutcome !== "delivered" && (await hasPriorRequesterDeliveryMirror(entry));
-      if (!isCleanupAttemptCurrent(runId, entry, cleanupGeneration)) {
-        await retireSupersededCleanupIfNeeded(runId, entry, cleanupGeneration);
-        return;
-      }
-      if (shouldCreditPriorDelivery) {
-        latestDeliveryError = undefined;
-      }
-      if (announceOutcome !== "delivered" && latestDeliveryError) {
-        ensureDeliveryState(entry).lastError = latestDeliveryError;
-      }
-      await finalizeSubagentCleanup(
-        runId,
-        cleanup,
-        shouldCreditPriorDelivery ? "delivered" : announceOutcome,
-        cleanupGeneration,
-      );
-    };
-
-    const announceParams: Parameters<RunSubagentAnnounceFlow>[0] = {
-      childSessionKey: pendingPayload.childSessionKey,
-      childRunId: pendingPayload.childRunId,
-      requesterSessionKey: pendingPayload.requesterSessionKey,
-      requesterOrigin,
-      requesterDisplayKey: pendingPayload.requesterDisplayKey,
-      task: pendingPayload.task,
-      timeoutMs: params.subagentAnnounceTimeoutMs,
-      cleanup: suppressSessionEffects ? "keep" : cleanup,
-      roundOneReply: entry.completion?.resultText ?? undefined,
-      terminalReply: pendingPayload.terminalReply,
-      fallbackReply: entry.completion?.fallbackResultText ?? undefined,
-      waitForCompletion: false,
-      startedAt: pendingPayload.startedAt,
-      endedAt: pendingPayload.endedAt,
-      label: pendingPayload.label,
-      outcome: pendingPayload.outcome,
-      spawnMode: pendingPayload.spawnMode,
-      expectsCompletionMessage: pendingPayload.expectsCompletionMessage,
-      wakeOnDescendantSettle: pendingPayload.wakeOnDescendantSettle === true,
-      suppressChildSessionEffects: suppressSessionEffects,
-      isChildSessionEffectsAllowed: childSessionEffectsAllowed,
-      isCompletionDeliveryAllowed: () => isCleanupAttemptCurrent(runId, entry, cleanupGeneration),
-      isCompletionOwnedByRequesterYield: () =>
-        entry.requesterTurnYielded === true ||
-        entry.requesterSettleWake?.requesterYieldBatch === true,
-      onBeforeDeleteChildSession:
-        cleanup === "delete"
-          ? () => {
-              if (!childSessionEffectsAllowed()) {
-                return false;
-              }
-              // Announce owns delete submission; fence late yields at the
-              // exact handoff instead of when cleanup merely starts.
-              entry.deleteCleanupDispatchedAt ??= Date.now();
-              params.persist(runId);
-              return true;
-            }
-          : undefined,
-      onDeliveryResult: (delivery) => {
-        if (!isCleanupAttemptCurrent(runId, entry, cleanupGeneration)) {
-          retireSupersededCleanupInBackground(runId, entry, cleanupGeneration);
+        const current = params.runs.get(args.runId);
+        if (
+          !current ||
+          current.cleanupCompletedAt ||
+          !isCleanupAttemptCurrent(context, params, args.runId, args.entry, args.cleanupGeneration)
+        ) {
           return;
         }
-        recordAnnounceDeliveryResult(entry, delivery);
-        if (delivery.delivered) {
-          const deliveryState = ensureDeliveryState(entry);
-          deliveryState.status = "delivered";
-          deliveryState.announcedAt = deliveryState.deliveredAt ?? Date.now();
-          deliveryState.lastError = undefined;
-          deliveryState.suspendedAt = undefined;
-          deliveryState.suspendedReason = undefined;
-          // Identified platform delivery precedes best-effort transcript
-          // mirroring; task ownership must become durable at that same edge.
-          params.persist(runId);
-          safeSetSubagentTaskDeliveryStatus({ entry, deliveryStatus: "delivered" });
-          latestDeliveryError = undefined;
-          return;
-        }
-        if (delivery.path === "none" && delivery.disposition !== "intentional_non_delivery") {
-          ensureDeliveryState(entry).lastDropReason = "sink_unavailable";
-        }
-        latestDeliveryError = formatAnnounceDeliveryError(delivery);
-        if (ensureDeliveryState(entry).lastError !== latestDeliveryError) {
-          ensureDeliveryState(entry).lastError = latestDeliveryError;
-          params.persist(runId);
-        }
-      },
-    };
-    runDetachedCleanupAttempt({
-      runId,
-      entry,
-      cleanupGeneration,
-      run: async () => {
-        let announceOutcome: SubagentAnnounceFlowOutcome = "retryable";
-        try {
-          announceOutcome = await params.runSubagentAnnounceFlow(announceParams);
-        } catch (error) {
-          defaultRuntime.log(
-            `[warn] Subagent announce flow failed during cleanup for run ${runId}: ${String(error)}`,
+        current.cleanupHandled = false;
+        params.resumedRuns.delete(args.runId);
+        params.persist(args.runId);
+        const failureCount = context.incrementCleanupFailureCount(current);
+        if (failureCount <= MAX_DETACHED_CLEANUP_RETRIES) {
+          scheduleResumeSubagentRun(
+            context,
+            params,
+            args.runId,
+            current,
+            resolveAnnounceRetryDelayMs(failureCount),
+            args.cleanupGeneration,
           );
         }
-        await finalizeAnnounceCleanup(announceOutcome);
-      },
+      }
+    }).catch((err: unknown) => {
+      defaultRuntime.log(
+        `[warn] subagent cleanup admission failed (${args.runId}): ${String(err)}`,
+      );
+      if (isGatewayRestartDraining()) {
+        scheduleResumeSubagentRun(
+          context,
+          params,
+          args.runId,
+          args.entry,
+          MIN_ANNOUNCE_RETRY_DELAY_MS,
+          args.cleanupGeneration,
+        );
+      }
     });
+  });
+}
+
+export function suspendPendingFinalDelivery(
+  context: SubagentLifecycleCleanupContext & SubagentLifecycleWakeContext,
+  params: SubagentLifecycleOptions,
+  args: {
+    runId: string;
+    entry: SubagentRunRecord;
+    reason: "expiry" | "permanent_failure";
+    error?: string;
+  },
+): void {
+  const previousEntry = structuredClone(args.entry);
+  markPendingFinalDelivery({
+    entry: args.entry,
+    error: args.error ?? getDeliveryLastError(args.entry) ?? args.reason,
+  });
+  const now = Date.now();
+  const delivery = ensureDeliveryState(args.entry);
+  delivery.status = "suspended";
+  delivery.suspendedAt ??= now;
+  delivery.suspendedReason = args.reason;
+  args.entry.cleanupHandled = false;
+  args.entry.wakeOnDescendantSettle = undefined;
+  const completion = ensureCompletionState(args.entry);
+  completion.fallbackResultText = undefined;
+  completion.fallbackCapturedAt = undefined;
+  params.resumedRuns.delete(args.runId);
+  safeSetSubagentTaskDeliveryStatus(params, context, {
+    entry: args.entry,
+    deliveryStatus: "failed",
+    deliveryError: getDeliveryLastError(args.entry) ?? args.reason,
+  });
+  safeMarkRequiredCompletionDeliveryBlocked(params, context, {
+    entry: args.entry,
+    reason: getDeliveryLastError(args.entry) ?? args.reason,
+  });
+  logAnnounceGiveUp(args.entry, args.reason);
+  markRequesterSettleWakePending(args.entry);
+  try {
+    params.persistOrThrow(args.runId);
+  } catch (error) {
+    const mutableEntry = args.entry as unknown as Record<string, unknown>;
+    for (const key of Object.keys(mutableEntry)) {
+      delete mutableEntry[key];
+    }
+    Object.assign(args.entry, previousEntry);
+    throw error;
+  }
+  // Suspension is terminal for automatic retries, so it settles this child
+  // for requester-drain purposes even though cleanup stays incomplete.
+  scheduleRequesterSettleWake(context, params, args.runId, args.entry);
+}
+
+export function beginSubagentCleanup(
+  context: SubagentLifecycleCleanupContext,
+  params: SubagentLifecycleOptions,
+  runId: string,
+): boolean {
+  const entry = params.runs.get(runId);
+  if (!entry || entry.cleanupCompletedAt || entry.cleanupHandled) {
+    return false;
+  }
+  entry.cleanupHandled = true;
+  context.bumpCleanupGeneration(entry);
+  params.persist(runId);
+  return true;
+}
+
+export async function retireSupersededCleanupIfNeeded(
+  context: SubagentLifecycleCleanupContext,
+  params: SubagentLifecycleOptions,
+  runId: string,
+  entry: SubagentRunRecord,
+  generation: number,
+): Promise<boolean> {
+  if (
+    params.runs.get(runId) !== entry ||
+    !context.isCleanupGeneration(entry, generation) ||
+    !context.newerGenerationOwnsSession(entry)
+  ) {
+    return false;
+  }
+  // Cleanup can yield to attachment, mirror, or announce work. A successor
+  // registered while it was suspended owns every session-scoped side effect.
+  await params.retireSupersededRun(runId, entry);
+  return true;
+}
+
+export function retireSupersededCleanupInBackground(
+  context: SubagentLifecycleCleanupContext,
+  params: SubagentLifecycleOptions,
+  runId: string,
+  entry: SubagentRunRecord,
+  generation: number,
+): void {
+  // Delivery callbacks are synchronous and may arrive after their announce
+  // attempt returns. Give the async retirement tail its own snapshot blocker.
+  void runWithGatewayIndependentRootWorkAdmission(async () => {
+    await retireSupersededCleanupIfNeeded(context, params, runId, entry, generation);
+  }).catch((error: unknown) => {
+    defaultRuntime.log(
+      `[warn] subagent superseded cleanup retirement failed (${runId}): ${String(error)}`,
+    );
+  });
+}
+
+export function isTerminalCallbackCurrent(
+  context: SubagentLifecycleCompletionContext,
+  params: SubagentLifecycleOptions,
+  runId: string,
+  entry: SubagentRunRecord,
+  generation: number,
+): boolean {
+  return (
+    params.runs.get(runId) === entry &&
+    entry.pauseReason !== "sessions_yield" &&
+    context.isTerminalGeneration(entry, generation)
+  );
+}
+
+export function isEndedHookOwnerCurrent(
+  context: SubagentLifecycleCleanupContext,
+  params: SubagentLifecycleOptions,
+  runId: string,
+  entry: SubagentRunRecord,
+): boolean {
+  const current = params.runs.get(runId);
+  return (current === undefined || current === entry) && !context.newerGenerationOwnsSession(entry);
+}
+
+export async function retireRunModeBundleMcpRuntime(
+  context: SubagentLifecycleCommonContext,
+  params: SubagentLifecycleOptions,
+  cleanupParams: { runId: string; entry: SubagentRunRecord; reason: string },
+): Promise<void> {
+  if (cleanupParams.entry.spawnMode === "session") {
+    return;
+  }
+  await retireSessionMcpRuntimeForSessionKey({
+    sessionKey: cleanupParams.entry.childSessionKey,
+    reason: cleanupParams.reason,
+    preserveActiveLeases: true,
+    onError: (error, sessionId) => {
+      params.warn("failed to retire subagent bundle MCP runtime", {
+        error: context.buildSafeLifecycleErrorMeta(error),
+        sessionId,
+        runId: context.maskRunId(cleanupParams.runId),
+        childSessionKey: context.maskSessionKey(cleanupParams.entry.childSessionKey),
+      });
+    },
+  });
+}
+
+export async function completeTerminalEffects(
+  context: SubagentLifecycleCompletionContext,
+  params: SubagentLifecycleOptions,
+  args: {
+    completeParams: SubagentCompletionRequest;
+    completionReason: SubagentLifecycleEndedReason;
+    entry: SubagentRunRecord;
+    mutated: boolean;
+    sessionSuperseded: boolean;
+    suppressSessionEffects: boolean;
+    terminalGeneration: number;
+    loadCleanupBrowserSessionsForLifecycleEnd(): Promise<
+      typeof cleanupBrowserSessionsForLifecycleEnd
+    >;
+  },
+): Promise<void> {
+  const { completeParams, completionReason, entry, mutated, terminalGeneration } = args;
+  let { sessionSuperseded, suppressSessionEffects } = args;
+  const isCurrentTerminalCallback = () =>
+    isTerminalCallbackCurrent(context, params, completeParams.runId, entry, terminalGeneration);
+  const isCurrentSessionEffectsOwner = () =>
+    isCurrentTerminalCallback() &&
+    !context.newerGenerationOwnsSession(entry) &&
+    !shouldSuppressSubagentRecoverySessionEffects(entry);
+  const refreshSessionEffectsSuppression = () => {
+    if (!shouldSuppressSubagentRecoverySessionEffects(entry)) {
+      return false;
+    }
+    suppressSessionEffects = true;
+    if (entry.execution.suppressSessionEffects !== true) {
+      const previousExecution = entry.execution;
+      entry.execution = { ...entry.execution, suppressSessionEffects: true };
+      try {
+        params.persistOrThrow(completeParams.runId);
+      } catch (error) {
+        entry.execution = previousExecution;
+        suppressSessionEffects = false;
+        throw error;
+      }
+    }
     return true;
   };
-
-  return {
-    completeCleanupBookkeeping,
-    finalizeResumedAnnounceGiveUp: finalizeAnnounceGiveUp,
-    retireRunModeBundleMcpRuntime,
-    startSubagentAnnounceCleanupFlow,
+  if (!isCurrentTerminalCallback()) {
+    return;
+  }
+  const retireSupersededSession = async (currentEntry: SubagentRunRecord) => {
+    if (completionReason !== SUBAGENT_ENDED_REASON_KILLED) {
+      await params.retireSupersededRun(completeParams.runId, currentEntry);
+    }
   };
+  sessionSuperseded ||= context.newerGenerationOwnsSession(entry);
+  if (sessionSuperseded) {
+    // This callback belongs to an older run that shared the session key.
+    // Update only its task projection; the newer generation owns all session effects.
+    await retireSupersededSession(entry);
+    return;
+  }
+  if (entry.collect) {
+    releaseSwarmRun(entry.schedulerSlotId ?? entry.runId);
+  }
+  refreshSessionEffectsSuppression();
+  const isProvisionalKill = entry.killReconciliation !== undefined;
+  // Record only the current, non-superseded callback with a committed outcome; the
+  // run-terminal dedupe key is first-write-wins, so a provisional/stale status here
+  // would permanently mislabel the signal-log terminal kind.
+  const outcomeStatus = entry.execution.outcome?.status;
+  if (
+    !suppressSessionEffects &&
+    !isProvisionalKill &&
+    outcomeStatus &&
+    outcomeStatus !== "unknown"
+  ) {
+    recordSubagentTerminalState({
+      childSessionKey: entry.childSessionKey,
+      runId: entry.runId,
+      requesterSessionKey: entry.requesterSessionKey,
+      outcomeStatus,
+    });
+  }
+
+  if (!suppressSessionEffects) {
+    try {
+      const assertSessionEffectsOwnerCurrent = () => {
+        if (!isCurrentSessionEffectsOwner()) {
+          throw new Error("subagent session-effects owner retired before session commit");
+        }
+      };
+      await persistSubagentSessionTiming(entry, {
+        // Recheck while patchSessionEntry owns its write lock so this old
+        // completion cannot commit after a synchronous ownership transfer.
+        isCurrentGeneration: isCurrentSessionEffectsOwner,
+        assertCommitAllowed: assertSessionEffectsOwnerCurrent,
+      });
+    } catch (err) {
+      params.warn("failed to persist subagent session timing", {
+        err,
+        runId: entry.runId,
+        childSessionKey: entry.childSessionKey,
+      });
+    }
+  }
+  refreshSessionEffectsSuppression();
+  if (!isCurrentTerminalCallback()) {
+    return;
+  }
+  if (context.newerGenerationOwnsSession(entry)) {
+    await retireSupersededSession(entry);
+    return;
+  }
+
+  const suppressedForSteerRestart = params.suppressAnnounceForSteerRestart(entry);
+  // Recovery persists its terminal state before draining this callback, so an
+  // unchanged row still needs its first session-status and progress events.
+  const shouldPublishTerminalStatus =
+    mutated ||
+    (completeParams.recoverInterrupted === true &&
+      !isProvisionalKill &&
+      !context.hasProgressEnded(entry));
+  if (shouldPublishTerminalStatus && !suppressedForSteerRestart && !suppressSessionEffects) {
+    emitSessionLifecycleEvent({
+      sessionKey: entry.childSessionKey,
+      reason: "subagent-status",
+      parentSessionKey: entry.requesterSessionKey,
+      label: entry.label,
+    });
+    // The enclosing steer/session-effects guard admits only the real terminal generation.
+    if (!isProvisionalKill && !context.hasProgressEnded(entry)) {
+      context.markProgressEnded(entry);
+      await params.emitSubagentProgressEndedForRun(entry);
+      refreshSessionEffectsSuppression();
+      if (!isCurrentTerminalCallback()) {
+        return;
+      }
+    }
+  }
+  const shouldEmitEndedHook =
+    !suppressedForSteerRestart &&
+    !isProvisionalKill &&
+    !suppressSessionEffects &&
+    params.shouldEmitEndedHookForRun({ entry, reason: completionReason });
+  const shouldDeferEndedHook =
+    shouldEmitEndedHook &&
+    completeParams.triggerCleanup &&
+    entry.expectsCompletionMessage === true &&
+    !suppressedForSteerRestart;
+  if (!shouldDeferEndedHook && shouldEmitEndedHook) {
+    await params.emitSubagentEndedHookForRun({
+      entry,
+      reason: completionReason,
+      sendFarewell: completeParams.sendFarewell,
+      accountId: completeParams.accountId,
+      isCurrent: isCurrentSessionEffectsOwner,
+    });
+    refreshSessionEffectsSuppression();
+    if (!isCurrentTerminalCallback()) {
+      return;
+    }
+    if (context.newerGenerationOwnsSession(entry)) {
+      await retireSupersededSession(entry);
+      return;
+    }
+  }
+
+  refreshSessionEffectsSuppression();
+  await completeTerminalCleanup(context, params, {
+    completeParams,
+    entry,
+    isProvisionalKill,
+    retireSupersededSession,
+    suppressedForSteerRestart,
+    suppressSessionEffects,
+    terminalGeneration,
+    loadCleanupBrowserSessionsForLifecycleEnd: () =>
+      args.loadCleanupBrowserSessionsForLifecycleEnd(),
+  });
+}
+
+export async function completeTerminalCleanup(
+  context: SubagentLifecycleCompletionContext,
+  params: SubagentLifecycleOptions,
+  args: {
+    completeParams: SubagentCompletionRequest;
+    entry: SubagentRunRecord;
+    isProvisionalKill: boolean;
+    retireSupersededSession: (entry: SubagentRunRecord) => Promise<void>;
+    suppressedForSteerRestart: boolean;
+    suppressSessionEffects: boolean;
+    terminalGeneration: number;
+    loadCleanupBrowserSessionsForLifecycleEnd(): Promise<
+      typeof cleanupBrowserSessionsForLifecycleEnd
+    >;
+  },
+): Promise<void> {
+  const {
+    completeParams,
+    entry,
+    isProvisionalKill,
+    retireSupersededSession,
+    suppressedForSteerRestart,
+    terminalGeneration,
+  } = args;
+  let { suppressSessionEffects } = args;
+  // Session cleanup belongs to the exact registry row and child generation.
+  // A replacement may reuse either the run id or the child session key.
+  const isSessionEffectsOwnerCurrent = () =>
+    isTerminalCallbackCurrent(context, params, completeParams.runId, entry, terminalGeneration) &&
+    !context.newerGenerationOwnsSession(entry);
+  const refreshSessionEffectsSuppression = () => {
+    if (
+      suppressSessionEffects ||
+      !isSessionEffectsOwnerCurrent() ||
+      !shouldSuppressSubagentRecoverySessionEffects(entry)
+    ) {
+      return suppressSessionEffects;
+    }
+    const previousExecution = entry.execution;
+    entry.execution = { ...previousExecution, suppressSessionEffects: true };
+    try {
+      params.persistOrThrow(completeParams.runId);
+    } catch (error) {
+      entry.execution = previousExecution;
+      throw error;
+    }
+    suppressSessionEffects = true;
+    return true;
+  };
+  if (!completeParams.triggerCleanup || suppressedForSteerRestart) {
+    return;
+  }
+  refreshSessionEffectsSuppression();
+  if (
+    !isTerminalCallbackCurrent(context, params, completeParams.runId, entry, terminalGeneration)
+  ) {
+    return;
+  }
+  if (context.newerGenerationOwnsSession(entry)) {
+    await retireSupersededSession(entry);
+    return;
+  }
+
+  // registerSubagentRun fires both an in-process listener and a gateway
+  // waitForSubagentCompletion RPC; both can reach this point for the same
+  // runId in embedded mode. Dedupe only the browser driver tab-close IPC
+  // with a sync check-then-set. The retire + announce tail below must still
+  // run for every caller, so a slow or held first browser cleanup cannot
+  // strand a duplicate caller's completion behind it.
+  if (!suppressSessionEffects && entry.browserCleanupDispatchedAt === undefined) {
+    let dispatchedBrowserCleanup = false;
+    let cleanupBrowserSessions: typeof cleanupBrowserSessionsForLifecycleEnd | undefined =
+      params.cleanupBrowserSessionsForLifecycleEnd;
+    try {
+      cleanupBrowserSessions ??= await args.loadCleanupBrowserSessionsForLifecycleEnd();
+    } catch (error) {
+      params.warn("failed to load browser cleanup for completed subagent", {
+        error: context.buildSafeLifecycleErrorMeta(error),
+        runId: context.maskRunId(completeParams.runId),
+        childSessionKey: context.maskSessionKey(entry.childSessionKey),
+      });
+    }
+    if (cleanupBrowserSessions) {
+      if (
+        !isTerminalCallbackCurrent(context, params, completeParams.runId, entry, terminalGeneration)
+      ) {
+        return;
+      }
+      if (context.newerGenerationOwnsSession(entry)) {
+        await retireSupersededSession(entry);
+        return;
+      }
+      if (refreshSessionEffectsSuppression()) {
+        return;
+      }
+      // Claim only when this caller is about to dispatch. A concurrent caller
+      // may have claimed while the lazy browser module was loading.
+      if (entry.browserCleanupDispatchedAt === undefined) {
+        entry.browserCleanupDispatchedAt = Date.now();
+        dispatchedBrowserCleanup = true;
+        try {
+          await cleanupBrowserSessions({
+            sessionKeys: [entry.childSessionKey],
+            onWarn: (msg) => params.warn(msg, { runId: entry.runId }),
+          });
+        } catch (error) {
+          params.warn("failed to cleanup browser sessions for completed subagent", {
+            error: context.buildSafeLifecycleErrorMeta(error),
+            runId: context.maskRunId(completeParams.runId),
+            childSessionKey: context.maskSessionKey(entry.childSessionKey),
+          });
+        }
+      }
+    }
+    if (dispatchedBrowserCleanup) {
+      if (
+        !isTerminalCallbackCurrent(context, params, completeParams.runId, entry, terminalGeneration)
+      ) {
+        return;
+      }
+      refreshSessionEffectsSuppression();
+      if (context.newerGenerationOwnsSession(entry)) {
+        await retireSupersededSession(entry);
+        return;
+      }
+    }
+  }
+
+  if (!suppressSessionEffects) {
+    if (
+      !isTerminalCallbackCurrent(context, params, completeParams.runId, entry, terminalGeneration)
+    ) {
+      return;
+    }
+    if (context.newerGenerationOwnsSession(entry)) {
+      await retireSupersededSession(entry);
+      return;
+    }
+    try {
+      await retireRunModeBundleMcpRuntime(context, params, {
+        runId: completeParams.runId,
+        entry,
+        reason: "subagent-run-complete",
+      });
+    } catch (error) {
+      params.warn("failed to retire subagent bundle MCP runtime after completion", {
+        error: context.buildSafeLifecycleErrorMeta(error),
+        runId: context.maskRunId(completeParams.runId),
+        childSessionKey: context.maskSessionKey(entry.childSessionKey),
+      });
+    }
+    if (
+      !isTerminalCallbackCurrent(context, params, completeParams.runId, entry, terminalGeneration)
+    ) {
+      return;
+    }
+    refreshSessionEffectsSuppression();
+    if (context.newerGenerationOwnsSession(entry)) {
+      await retireSupersededSession(entry);
+      return;
+    }
+  }
+
+  if (isProvisionalKill) {
+    // Browser and MCP resources can close immediately, but completion delivery
+    // waits for the provider result or the killed tombstone reconciliation.
+    return;
+  }
+
+  refreshSessionEffectsSuppression();
+  context.startSubagentAnnounceCleanupFlow(completeParams.runId, entry);
 }
