+++ src/agents/subagents/registry/subagent-registry-lifecycle-delivery.ts
@@ -31,475 +31,471 @@ import {
 import type { SubagentLifecycleEndedReason } from "./subagent-lifecycle-events.js";
 import { resolveFinalizedSubagentTaskState } from "./subagent-registry-completion.js";
 import { capFrozenResultText } from "./subagent-registry-helpers.js";
-import type { createSubagentRegistryLifecycleCommon } from "./subagent-registry-lifecycle-common.js";
 import type {
-  SubagentRegistryLifecycleParams,
-  SubagentRegistryLifecycleState,
-} from "./subagent-registry-lifecycle-contracts.js";
+  SubagentLifecycleCommonContext,
+  SubagentLifecycleOptions,
+} from "./subagent-registry-lifecycle.js";
 import type { PendingFinalDeliveryPayload, SubagentRunRecord } from "./subagent-registry.types.js";
 import { compareSubagentRunGeneration } from "./subagent-run-generation.js";
 
 const DELIVERY_MIRROR_HISTORY_MAX_CHARS = 128 * 1024;
 
-export function createSubagentRegistryLifecycleDelivery(
-  params: SubagentRegistryLifecycleParams,
-  _state: SubagentRegistryLifecycleState,
-  common: ReturnType<typeof createSubagentRegistryLifecycleCommon>,
-) {
-  const { newerGenerationOwnsSession, buildSafeLifecycleErrorMeta, maskRunId, maskSessionKey } =
-    common;
+export const formatAnnounceDeliveryError = (delivery: SubagentAnnounceDeliveryResult): string => {
+  const errors = [
+    delivery.error,
+    delivery.reason,
+    ...(delivery.phases ?? []).map((phase) =>
+      phase.error ? `${phase.phase}: ${phase.error}` : undefined,
+    ),
+  ]
+    .map((value) => value?.trim())
+    .filter((value): value is string => Boolean(value));
+  return errors.length > 0
+    ? uniqueStrings(errors).join("; ")
+    : `delivery path ${delivery.path} did not complete`;
+};
 
-  const formatAnnounceDeliveryError = (delivery: SubagentAnnounceDeliveryResult): string => {
-    const errors = [
-      delivery.error,
-      delivery.reason,
-      ...(delivery.phases ?? []).map((phase) =>
-        phase.error ? `${phase.phase}: ${phase.error}` : undefined,
-      ),
-    ]
-      .map((value) => value?.trim())
-      .filter((value): value is string => Boolean(value));
-    return errors.length > 0
-      ? uniqueStrings(errors).join("; ")
-      : `delivery path ${delivery.path} did not complete`;
-  };
+export const recordAnnounceDeliveryResult = (
+  entry: SubagentRunRecord,
+  delivery: SubagentAnnounceDeliveryResult,
+) => {
+  const deliveryState = ensureDeliveryState(entry);
+  if (typeof delivery.enqueuedAt === "number") {
+    deliveryState.enqueuedAt ??= delivery.enqueuedAt;
+  }
+  if (delivery.delivered) {
+    const deliveredAt =
+      typeof delivery.deliveredAt === "number" ? delivery.deliveredAt : Date.now();
+    deliveryState.deliveredAt = deliveredAt;
+    deliveryState.lastDropReason = undefined;
+  }
+  deliveryState.disposition =
+    delivery.disposition ?? (delivery.delivered ? "delivered" : "retryable");
+};
 
-  const recordAnnounceDeliveryResult = (
-    entry: SubagentRunRecord,
-    delivery: SubagentAnnounceDeliveryResult,
-  ) => {
-    const deliveryState = ensureDeliveryState(entry);
-    if (typeof delivery.enqueuedAt === "number") {
-      deliveryState.enqueuedAt ??= delivery.enqueuedAt;
-    }
-    if (delivery.delivered) {
-      const deliveredAt =
-        typeof delivery.deliveredAt === "number" ? delivery.deliveredAt : Date.now();
-      deliveryState.deliveredAt = deliveredAt;
-      deliveryState.lastDropReason = undefined;
-    }
-    deliveryState.disposition =
-      delivery.disposition ?? (delivery.delivered ? "delivered" : "retryable");
-  };
-
-  const hasPriorRequesterDeliveryMirror = async (entry: SubagentRunRecord): Promise<boolean> => {
-    const completion = ensureCompletionState(entry);
-    const expectedText = extractTextFromChatContent(completion.resultText, { joinWith: "" });
-    if (entry.expectsCompletionMessage !== true || expectedText == null) {
-      return false;
-    }
-    const mirrorNotBefore = entry.execution.startedAt ?? entry.createdAt;
-    const mirrorNotAfter = Date.now() + 30_000;
-    const expectedIdempotencyKey = buildAnnounceIdempotencyKey(
-      buildAnnounceIdFromChildRun({
-        childSessionKey: entry.childSessionKey,
-        childRunId: entry.runId,
-      }),
-    );
-    const isExpectedMirrorIdempotencyKey = (value: unknown): boolean =>
-      typeof value === "string" &&
-      (value === expectedIdempotencyKey ||
-        value.startsWith(`${expectedIdempotencyKey}:internal-source-reply:`) ||
-        value.startsWith(`${expectedIdempotencyKey}:message-tool:internal-source-reply:`) ||
-        value.startsWith(`${entry.runId}:message-tool:`) ||
-        value.startsWith(`${entry.runId}:internal-source-reply:`));
-    try {
-      const history = await params.callGateway<{
-        messages?: unknown[];
-      }>({
-        method: "chat.history",
-        params: {
-          sessionKey: entry.requesterSessionKey,
-          limit: 25,
-          maxChars: DELIVERY_MIRROR_HISTORY_MAX_CHARS,
-        },
-        timeoutMs: 5_000,
-      });
-      const mirror = history.messages?.find((message) => {
-        if (!message || typeof message !== "object") {
-          return false;
-        }
-        const record = message as Record<string, unknown>;
-        const timestamp = record.timestamp;
-        if (
-          typeof timestamp !== "number" ||
-          !Number.isFinite(timestamp) ||
-          timestamp < mirrorNotBefore ||
-          timestamp > mirrorNotAfter ||
-          !isExpectedMirrorIdempotencyKey(record.idempotencyKey)
-        ) {
-          return false;
-        }
-        const text = extractTextFromChatContent(record.content, { joinWith: "" });
-        return (
-          record.role === "assistant" &&
-          record.provider === "openclaw" &&
-          record.model === "delivery-mirror" &&
-          text === expectedText
-        );
-      });
-      if (mirror) {
-        ensureDeliveryState(entry).deliveredAt = (mirror as { timestamp: number }).timestamp;
+export const hasPriorRequesterDeliveryMirror = async (
+  params: SubagentLifecycleOptions,
+  entry: SubagentRunRecord,
+): Promise<boolean> => {
+  const completion = ensureCompletionState(entry);
+  const expectedText = extractTextFromChatContent(completion.resultText, { joinWith: "" });
+  if (entry.expectsCompletionMessage !== true || expectedText == null) {
+    return false;
+  }
+  const mirrorNotBefore = entry.execution.startedAt ?? entry.createdAt;
+  const mirrorNotAfter = Date.now() + 30_000;
+  const expectedIdempotencyKey = buildAnnounceIdempotencyKey(
+    buildAnnounceIdFromChildRun({
+      childSessionKey: entry.childSessionKey,
+      childRunId: entry.runId,
+    }),
+  );
+  const isExpectedMirrorIdempotencyKey = (value: unknown): boolean =>
+    typeof value === "string" &&
+    (value === expectedIdempotencyKey ||
+      value.startsWith(`${expectedIdempotencyKey}:internal-source-reply:`) ||
+      value.startsWith(`${expectedIdempotencyKey}:message-tool:internal-source-reply:`) ||
+      value.startsWith(`${entry.runId}:message-tool:`) ||
+      value.startsWith(`${entry.runId}:internal-source-reply:`));
+  try {
+    const history = await params.callGateway<{
+      messages?: unknown[];
+    }>({
+      method: "chat.history",
+      params: {
+        sessionKey: entry.requesterSessionKey,
+        limit: 25,
+        maxChars: DELIVERY_MIRROR_HISTORY_MAX_CHARS,
+      },
+      timeoutMs: 5_000,
+    });
+    const mirror = history.messages?.find((message) => {
+      if (!message || typeof message !== "object") {
+        return false;
       }
-      return Boolean(mirror);
-    } catch {
-      return false;
+      const record = message as Record<string, unknown>;
+      const timestamp = record.timestamp;
+      if (
+        typeof timestamp !== "number" ||
+        !Number.isFinite(timestamp) ||
+        timestamp < mirrorNotBefore ||
+        timestamp > mirrorNotAfter ||
+        !isExpectedMirrorIdempotencyKey(record.idempotencyKey)
+      ) {
+        return false;
+      }
+      const text = extractTextFromChatContent(record.content, { joinWith: "" });
+      return (
+        record.role === "assistant" &&
+        record.provider === "openclaw" &&
+        record.model === "delivery-mirror" &&
+        text === expectedText
+      );
+    });
+    if (mirror) {
+      ensureDeliveryState(entry).deliveredAt = (mirror as { timestamp: number }).timestamp;
     }
-  };
+    return Boolean(mirror);
+  } catch {
+    return false;
+  }
+};
 
-  const resolveSubagentTaskTarget = (
-    entry: SubagentRunRecord,
-    resolution = params.resolveSubagentTask(entry),
-  ) => {
-    const durableTaskRunId = entry.taskRunId ?? entry.runId;
-    return {
-      runId:
-        resolution.lookup === "available"
-          ? (resolution.task?.runId ?? durableTaskRunId)
-          : durableTaskRunId,
-      sessionKey:
-        resolution.lookup === "available"
-          ? (resolution.task?.childSessionKey ?? entry.childSessionKey)
-          : entry.childSessionKey,
-    };
+const resolveSubagentTaskTarget = (
+  params: SubagentLifecycleOptions,
+  entry: SubagentRunRecord,
+  resolution = params.resolveSubagentTask(entry),
+) => {
+  const durableTaskRunId = entry.taskRunId ?? entry.runId;
+  return {
+    runId:
+      resolution.lookup === "available"
+        ? (resolution.task?.runId ?? durableTaskRunId)
+        : durableTaskRunId,
+    sessionKey:
+      resolution.lookup === "available"
+        ? (resolution.task?.childSessionKey ?? entry.childSessionKey)
+        : entry.childSessionKey,
   };
+};
 
-  const safeSetSubagentTaskDeliveryStatus = (args: {
+export const safeSetSubagentTaskDeliveryStatus = (
+  params: SubagentLifecycleOptions,
+  context: SubagentLifecycleCommonContext,
+  args: {
     entry: SubagentRunRecord;
     deliveryStatus: Extract<TaskDeliveryStatus, "pending" | "delivered" | "failed">;
     deliveryError?: string;
-  }) => {
-    const target = resolveSubagentTaskTarget(args.entry);
-    try {
-      setDetachedTaskDeliveryStatusByRunId({
-        runId: target.runId,
-        runtime: "subagent",
-        sessionKey: target.sessionKey,
-        deliveryStatus: args.deliveryStatus,
-        error: args.deliveryStatus === "failed" ? args.deliveryError : undefined,
-      });
-    } catch (err) {
-      params.warn("failed to update subagent background task delivery state", {
-        error: buildSafeLifecycleErrorMeta(err),
-        runId: maskRunId(target.runId),
-        childSessionKey: maskSessionKey(target.sessionKey),
-        deliveryStatus: args.deliveryStatus,
-      });
-    }
-  };
+  },
+) => {
+  const target = resolveSubagentTaskTarget(params, args.entry);
+  try {
+    setDetachedTaskDeliveryStatusByRunId({
+      runId: target.runId,
+      runtime: "subagent",
+      sessionKey: target.sessionKey,
+      deliveryStatus: args.deliveryStatus,
+      error: args.deliveryStatus === "failed" ? args.deliveryError : undefined,
+    });
+  } catch (err) {
+    params.warn("failed to update subagent background task delivery state", {
+      error: context.buildSafeLifecycleErrorMeta(err),
+      runId: context.maskRunId(target.runId),
+      childSessionKey: context.maskSessionKey(target.sessionKey),
+      deliveryStatus: args.deliveryStatus,
+    });
+  }
+};
 
-  const safeFinalizeSubagentTaskRun = (args: {
+export const safeFinalizeSubagentTaskRun = (
+  params: SubagentLifecycleOptions,
+  context: SubagentLifecycleCommonContext,
+  args: {
     entry: SubagentRunRecord;
     outcome: SubagentRunOutcome;
     taskResolution?: DetachedTaskFindResult;
-  }): ReturnType<typeof completeTaskRunByRunId> => {
-    const terminal = resolveFinalizedSubagentTaskState(args.entry);
-    if (!terminal) {
-      return [];
-    }
-    const target = resolveSubagentTaskTarget(args.entry, args.taskResolution);
-    const { status, error, terminalOutcome, ...details } = terminal;
-    const suppressDelivery = args.entry.suppressCompletionDelivery === true;
-    try {
-      if (status === "succeeded") {
-        return completeTaskRunByRunId({
-          runId: target.runId,
-          runtime: "subagent",
-          sessionKey: target.sessionKey,
-          ...details,
-          terminalOutcome,
-          suppressDelivery,
-        });
-      }
-      return failTaskRunByRunId({
+  },
+): ReturnType<typeof completeTaskRunByRunId> => {
+  const terminal = resolveFinalizedSubagentTaskState(args.entry);
+  if (!terminal) {
+    return [];
+  }
+  const target = resolveSubagentTaskTarget(params, args.entry, args.taskResolution);
+  const { status, error, terminalOutcome, ...details } = terminal;
+  const suppressDelivery = args.entry.suppressCompletionDelivery === true;
+  try {
+    if (status === "succeeded") {
+      return completeTaskRunByRunId({
         runId: target.runId,
         runtime: "subagent",
         sessionKey: target.sessionKey,
         ...details,
-        status,
-        error,
+        terminalOutcome,
         suppressDelivery,
       });
-    } catch (err) {
-      params.warn("failed to finalize subagent background task state", {
-        error: buildSafeLifecycleErrorMeta(err),
-        runId: maskRunId(args.entry.runId),
-        childSessionKey: maskSessionKey(args.entry.childSessionKey),
-        outcomeStatus: args.outcome.status,
-      });
-      return [];
     }
-  };
+    return failTaskRunByRunId({
+      runId: target.runId,
+      runtime: "subagent",
+      sessionKey: target.sessionKey,
+      ...details,
+      status,
+      error,
+      suppressDelivery,
+    });
+  } catch (err) {
+    params.warn("failed to finalize subagent background task state", {
+      error: context.buildSafeLifecycleErrorMeta(err),
+      runId: context.maskRunId(args.entry.runId),
+      childSessionKey: context.maskSessionKey(args.entry.childSessionKey),
+      outcomeStatus: args.outcome.status,
+    });
+    return [];
+  }
+};
 
-  const safeMarkRequiredCompletionDeliveryBlocked = (args: {
+export const safeMarkRequiredCompletionDeliveryBlocked = (
+  params: SubagentLifecycleOptions,
+  context: SubagentLifecycleCommonContext,
+  args: {
     entry: SubagentRunRecord;
     reason?: string;
-  }) => {
-    if (
-      args.entry.expectsCompletionMessage !== true ||
-      args.entry.execution.outcome?.status !== "ok"
-    ) {
-      return;
-    }
-    const endedAt = args.entry.execution.endedAt ?? Date.now();
-    const terminalResult = resolveRequiredCompletionDeliveryFailureTerminalResult(args.reason);
-    const target = resolveSubagentTaskTarget(args.entry);
-    try {
-      completeTaskRunByRunId({
-        runId: target.runId,
-        runtime: "subagent",
-        sessionKey: target.sessionKey,
-        endedAt,
-        lastEventAt: Date.now(),
-        progressSummary: resolveSubagentCompletionResultText(args.entry),
-        terminalSummary: terminalResult.terminalSummary,
-        terminalOutcome: terminalResult.terminalOutcome,
-      });
-    } catch (err) {
-      params.warn("failed to mark subagent completion delivery blocked", {
-        error: buildSafeLifecycleErrorMeta(err),
-        runId: maskRunId(args.entry.runId),
-        childSessionKey: maskSessionKey(args.entry.childSessionKey),
-      });
-    }
-  };
+  },
+) => {
+  if (
+    args.entry.expectsCompletionMessage !== true ||
+    args.entry.execution.outcome?.status !== "ok"
+  ) {
+    return;
+  }
+  const endedAt = args.entry.execution.endedAt ?? Date.now();
+  const terminalResult = resolveRequiredCompletionDeliveryFailureTerminalResult(args.reason);
+  const target = resolveSubagentTaskTarget(params, args.entry);
+  try {
+    completeTaskRunByRunId({
+      runId: target.runId,
+      runtime: "subagent",
+      sessionKey: target.sessionKey,
+      endedAt,
+      lastEventAt: Date.now(),
+      progressSummary: resolveSubagentCompletionResultText(args.entry),
+      terminalSummary: terminalResult.terminalSummary,
+      terminalOutcome: terminalResult.terminalOutcome,
+    });
+  } catch (err) {
+    params.warn("failed to mark subagent completion delivery blocked", {
+      error: context.buildSafeLifecycleErrorMeta(err),
+      runId: context.maskRunId(args.entry.runId),
+      childSessionKey: context.maskSessionKey(args.entry.childSessionKey),
+    });
+  }
+};
 
-  const freezeRunResultAtCompletion = async (
-    entry: SubagentRunRecord,
-    outcome: SubagentRunOutcome,
-  ): Promise<boolean> => {
-    if (ensureCompletionState(entry).resultText !== undefined) {
-      return false;
-    }
-    if (outcome.status === "error") {
-      const completion = ensureCompletionState(entry);
-      completion.resultText = null;
-      completion.capturedAt = Date.now();
-      return true;
-    }
-    let resultText: string | null;
-    try {
-      const transcriptTarget = entry.execution.transcriptTarget;
-      const agentId =
-        transcriptTarget?.agentId ?? resolveAgentIdFromSessionKey(entry.childSessionKey);
-      const sessionKey = transcriptTarget?.sessionKey ?? entry.childSessionKey;
-      const configuredStorePath = agentId
-        ? (transcriptTarget?.storePath ??
-          resolveStorePath(params.getRuntimeConfig().session?.store, { agentId }))
-        : undefined;
-      const storePath = configuredStorePath
-        ? resolveSessionStorePathForScope({
-            agentId,
-            sessionKey,
-            storePath: configuredStorePath,
-          })
-        : undefined;
-      const sessionId =
-        transcriptTarget?.sessionId ??
-        (agentId && storePath
-          ? loadSessionEntryReadOnly({ agentId, sessionKey, storePath })?.sessionId
-          : undefined);
-      const sessionTarget: SessionTranscriptRuntimeTarget | undefined =
-        agentId && sessionId && storePath
-          ? { agentId, sessionId, sessionKey, storePath }
-          : undefined;
-      const captured = await params.captureSubagentCompletionReply(entry.childSessionKey, {
-        waitForReply: entry.expectsCompletionMessage === true,
-        outcome,
-        ...(sessionTarget ? { sessionTarget } : {}),
-      });
-      resultText = captured?.trim() ? capFrozenResultText(captured) : null;
-    } catch {
-      resultText = null;
-    }
-    const liveEntry = params.runs.get(entry.runId);
-    if (
-      entry.pauseReason === "sessions_yield" ||
-      liveEntry?.pauseReason === "sessions_yield" ||
-      newerGenerationOwnsSession(entry)
-    ) {
-      return false;
-    }
+export const freezeRunResultAtCompletion = async (
+  params: SubagentLifecycleOptions,
+  context: SubagentLifecycleCommonContext,
+  entry: SubagentRunRecord,
+  outcome: SubagentRunOutcome,
+): Promise<boolean> => {
+  if (ensureCompletionState(entry).resultText !== undefined) {
+    return false;
+  }
+  if (outcome.status === "error") {
     const completion = ensureCompletionState(entry);
-    if (completion.resultText !== undefined) {
-      return false;
-    }
-    completion.resultText = resultText;
+    completion.resultText = null;
     completion.capturedAt = Date.now();
     return true;
-  };
+  }
+  let resultText: string | null;
+  try {
+    const transcriptTarget = entry.execution.transcriptTarget;
+    const agentId =
+      transcriptTarget?.agentId ?? resolveAgentIdFromSessionKey(entry.childSessionKey);
+    const sessionKey = transcriptTarget?.sessionKey ?? entry.childSessionKey;
+    const configuredStorePath = agentId
+      ? (transcriptTarget?.storePath ??
+        resolveStorePath(params.getRuntimeConfig().session?.store, { agentId }))
+      : undefined;
+    const storePath = configuredStorePath
+      ? resolveSessionStorePathForScope({
+          agentId,
+          sessionKey,
+          storePath: configuredStorePath,
+        })
+      : undefined;
+    const sessionId =
+      transcriptTarget?.sessionId ??
+      (agentId && storePath
+        ? loadSessionEntryReadOnly({ agentId, sessionKey, storePath })?.sessionId
+        : undefined);
+    const sessionTarget: SessionTranscriptRuntimeTarget | undefined =
+      agentId && sessionId && storePath ? { agentId, sessionId, sessionKey, storePath } : undefined;
+    const captured = await params.captureSubagentCompletionReply(entry.childSessionKey, {
+      waitForReply: entry.expectsCompletionMessage === true,
+      outcome,
+      ...(sessionTarget ? { sessionTarget } : {}),
+    });
+    resultText = captured?.trim() ? capFrozenResultText(captured) : null;
+  } catch {
+    resultText = null;
+  }
+  const liveEntry = params.runs.get(entry.runId);
+  if (
+    entry.pauseReason === "sessions_yield" ||
+    liveEntry?.pauseReason === "sessions_yield" ||
+    context.newerGenerationOwnsSession(entry)
+  ) {
+    return false;
+  }
+  const completion = ensureCompletionState(entry);
+  if (completion.resultText !== undefined) {
+    return false;
+  }
+  completion.resultText = resultText;
+  completion.capturedAt = Date.now();
+  return true;
+};
 
-  const listPendingCompletionRunsForSession = (sessionKey: string): SubagentRunRecord[] => {
-    const key = sessionKey.trim();
-    if (!key) {
-      return [];
+const listPendingCompletionRunsForSession = (
+  params: SubagentLifecycleOptions,
+  sessionKey: string,
+): SubagentRunRecord[] => {
+  const key = sessionKey.trim();
+  if (!key) {
+    return [];
+  }
+  const out: SubagentRunRecord[] = [];
+  for (const entry of params.runs.values()) {
+    if (entry.childSessionKey !== key) {
+      continue;
     }
-    const out: SubagentRunRecord[] = [];
-    for (const entry of params.runs.values()) {
-      if (entry.childSessionKey !== key) {
-        continue;
-      }
-      if (entry.expectsCompletionMessage !== true) {
-        continue;
-      }
-      if (typeof entry.execution.endedAt !== "number") {
-        continue;
-      }
-      if (typeof entry.cleanupCompletedAt === "number") {
-        continue;
-      }
-      // A paused row's result was deliberately cleared when it yielded; the text
-      // now in its session belongs to whatever turn runs next, not to the paused
-      // work. Refreezing it here would announce a stranger's output as this run's
-      // completion once the row finally settles.
-      if (entry.pauseReason === "sessions_yield") {
-        continue;
-      }
-      out.push(entry);
+    if (entry.expectsCompletionMessage !== true) {
+      continue;
     }
-    return out;
-  };
-
-  const refreshFrozenResultFromSession = async (sessionKey: string): Promise<boolean> => {
-    const candidates = listPendingCompletionRunsForSession(sessionKey).filter(
-      (entry) => entry.execution.outcome?.status !== "error",
-    );
-    const entry = candidates.toSorted(compareSubagentRunGeneration).at(-1);
-    if (!entry || newerGenerationOwnsSession(entry)) {
-      return false;
+    if (typeof entry.execution.endedAt !== "number") {
+      continue;
     }
-    const generation = entry.generation;
-
-    let captured: string | undefined;
-    try {
-      captured = await params.captureSubagentCompletionReply(sessionKey);
-    } catch {
-      return false;
+    if (typeof entry.cleanupCompletedAt === "number") {
+      continue;
     }
-    const trimmed = captured?.trim();
-    if (!trimmed || isSilentAgentReplyText(trimmed)) {
-      return false;
+    // A paused row's result was deliberately cleared when it yielded; the text
+    // now in its session belongs to whatever turn runs next, not to the paused
+    // work. Refreezing it here would announce a stranger's output as this run's
+    // completion once the row finally settles.
+    if (entry.pauseReason === "sessions_yield") {
+      continue;
     }
-    // Reply capture yields while registration can transfer session ownership.
-    // Only the exact row and generation that started capture may commit its text.
-    if (
-      params.runs.get(entry.runId) !== entry ||
-      entry.generation !== generation ||
-      newerGenerationOwnsSession(entry)
-    ) {
-      return false;
-    }
+    out.push(entry);
+  }
+  return out;
+};
 
-    const nextFrozen = capFrozenResultText(trimmed);
-    const completion = ensureCompletionState(entry);
-    if (completion.resultText === nextFrozen) {
-      return false;
-    }
-    completion.resultText = nextFrozen;
-    completion.capturedAt = Date.now();
-    params.persist(entry.runId);
-    return true;
-  };
+export const refreshFrozenResultFromSession = async (
+  params: SubagentLifecycleOptions,
+  context: SubagentLifecycleCommonContext,
+  sessionKey: string,
+): Promise<boolean> => {
+  const candidates = listPendingCompletionRunsForSession(params, sessionKey).filter(
+    (entry) => entry.execution.outcome?.status !== "error",
+  );
+  const entry = candidates.toSorted(compareSubagentRunGeneration).at(-1);
+  if (!entry || context.newerGenerationOwnsSession(entry)) {
+    return false;
+  }
+  const generation = entry.generation;
 
-  const emitCompletionEndedHookIfNeeded = async (
-    entry: SubagentRunRecord,
-    reason: SubagentLifecycleEndedReason,
-    isCurrent?: () => boolean,
-  ) => {
-    if (params.shouldEmitEndedHookForRun({ entry, reason })) {
-      await params.emitSubagentEndedHookForRun({
-        entry,
-        reason,
-        sendFarewell: true,
-        isCurrent,
-      });
-    }
-  };
+  let captured: string | undefined;
+  try {
+    captured = await params.captureSubagentCompletionReply(sessionKey);
+  } catch {
+    return false;
+  }
+  const trimmed = captured?.trim();
+  if (!trimmed || isSilentAgentReplyText(trimmed)) {
+    return false;
+  }
+  // Reply capture yields while registration can transfer session ownership.
+  // Only the exact row and generation that started capture may commit its text.
+  if (
+    params.runs.get(entry.runId) !== entry ||
+    entry.generation !== generation ||
+    context.newerGenerationOwnsSession(entry)
+  ) {
+    return false;
+  }
 
-  const clearPendingFinalDelivery = (entry: SubagentRunRecord) => {
-    const delivery = ensureDeliveryState(entry);
-    delivery.payload = undefined;
-    delivery.createdAt = undefined;
-    delivery.lastAttemptAt = undefined;
-    delivery.attemptCount = undefined;
-    delivery.lastError = undefined;
-    delivery.suspendedAt = undefined;
-    delivery.suspendedReason = undefined;
-    if (delivery.status !== "delivered" && delivery.status !== "failed") {
-      clearDeliveryState(entry);
-    }
-  };
+  const nextFrozen = capFrozenResultText(trimmed);
+  const completion = ensureCompletionState(entry);
+  if (completion.resultText === nextFrozen) {
+    return false;
+  }
+  completion.resultText = nextFrozen;
+  completion.capturedAt = Date.now();
+  params.persist(entry.runId);
+  return true;
+};
 
-  const loadPendingFinalDeliveryPayload = (
-    entry: SubagentRunRecord,
-  ): PendingFinalDeliveryPayload => {
-    return {
-      requesterSessionKey:
-        entry.delivery?.payload?.requesterSessionKey ?? entry.requesterSessionKey,
-      requesterOrigin: entry.delivery?.payload?.requesterOrigin ?? entry.requesterOrigin,
-      requesterDisplayKey:
-        entry.delivery?.payload?.requesterDisplayKey ?? entry.requesterDisplayKey,
-      childSessionKey: entry.delivery?.payload?.childSessionKey ?? entry.childSessionKey,
-      childRunId: entry.delivery?.payload?.childRunId ?? entry.runId,
-      task: entry.delivery?.payload?.task ?? entry.task,
-      label: entry.delivery?.payload?.label ?? entry.label,
-      startedAt: entry.delivery?.payload?.startedAt ?? entry.execution.startedAt,
-      endedAt: entry.delivery?.payload?.endedAt ?? entry.execution.endedAt,
-      outcome: entry.delivery?.payload?.outcome ?? entry.execution.outcome,
-      expectsCompletionMessage:
-        entry.delivery?.payload?.expectsCompletionMessage ?? entry.expectsCompletionMessage,
-      spawnMode: entry.delivery?.payload?.spawnMode ?? entry.spawnMode,
-      wakeOnDescendantSettle:
-        entry.delivery?.payload?.wakeOnDescendantSettle ?? entry.wakeOnDescendantSettle,
-      terminalReply: entry.delivery?.payload?.terminalReply ?? entry.completion?.terminalReply,
-    };
-  };
+export const emitCompletionEndedHookIfNeeded = async (
+  params: SubagentLifecycleOptions,
+  entry: SubagentRunRecord,
+  reason: SubagentLifecycleEndedReason,
+  isCurrent?: () => boolean,
+) => {
+  if (params.shouldEmitEndedHookForRun({ entry, reason })) {
+    await params.emitSubagentEndedHookForRun({
+      entry,
+      reason,
+      sendFarewell: true,
+      isCurrent,
+    });
+  }
+};
 
-  const markPendingFinalDelivery = (args: { entry: SubagentRunRecord; error?: string }) => {
-    const now = Date.now();
-    const payload: PendingFinalDeliveryPayload = loadPendingFinalDeliveryPayload(args.entry);
+export const clearPendingFinalDelivery = (entry: SubagentRunRecord) => {
+  const delivery = ensureDeliveryState(entry);
+  delivery.payload = undefined;
+  delivery.createdAt = undefined;
+  delivery.lastAttemptAt = undefined;
+  delivery.attemptCount = undefined;
+  delivery.lastError = undefined;
+  delivery.suspendedAt = undefined;
+  delivery.suspendedReason = undefined;
+  if (delivery.status !== "delivered" && delivery.status !== "failed") {
+    clearDeliveryState(entry);
+  }
+};
 
-    const delivery = ensureDeliveryState(args.entry);
-    delivery.status = "pending";
-    delivery.createdAt ??= now;
-    delivery.lastAttemptAt = now;
-    delivery.attemptCount = (delivery.attemptCount ?? 0) + 1;
-    delivery.lastError = args.error ?? null;
-    delivery.payload = payload;
-  };
-
-  const refreshPendingFinalDeliveryPayload = (entry: SubagentRunRecord): boolean => {
-    const delivery = entry.delivery;
-    if (
-      !delivery?.payload ||
-      delivery.status === "delivered" ||
-      typeof delivery.announcedAt === "number"
-    ) {
-      return false;
-    }
-    delivery.payload = {
-      ...delivery.payload,
-      startedAt: entry.execution.startedAt,
-      endedAt: entry.execution.endedAt,
-      outcome: entry.execution.outcome,
-      terminalReply: entry.completion?.terminalReply,
-    };
-    return true;
-  };
-
+export const loadPendingFinalDeliveryPayload = (
+  entry: SubagentRunRecord,
+): PendingFinalDeliveryPayload => {
   return {
-    clearPendingFinalDelivery,
-    emitCompletionEndedHookIfNeeded,
-    formatAnnounceDeliveryError,
-    freezeRunResultAtCompletion,
-    hasPriorRequesterDeliveryMirror,
-    loadPendingFinalDeliveryPayload,
-    markPendingFinalDelivery,
-    recordAnnounceDeliveryResult,
-    refreshFrozenResultFromSession,
-    refreshPendingFinalDeliveryPayload,
-    safeFinalizeSubagentTaskRun,
-    safeMarkRequiredCompletionDeliveryBlocked,
-    safeSetSubagentTaskDeliveryStatus,
+    requesterSessionKey: entry.delivery?.payload?.requesterSessionKey ?? entry.requesterSessionKey,
+    requesterOrigin: entry.delivery?.payload?.requesterOrigin ?? entry.requesterOrigin,
+    requesterDisplayKey: entry.delivery?.payload?.requesterDisplayKey ?? entry.requesterDisplayKey,
+    childSessionKey: entry.delivery?.payload?.childSessionKey ?? entry.childSessionKey,
+    childRunId: entry.delivery?.payload?.childRunId ?? entry.runId,
+    task: entry.delivery?.payload?.task ?? entry.task,
+    label: entry.delivery?.payload?.label ?? entry.label,
+    startedAt: entry.delivery?.payload?.startedAt ?? entry.execution.startedAt,
+    endedAt: entry.delivery?.payload?.endedAt ?? entry.execution.endedAt,
+    outcome: entry.delivery?.payload?.outcome ?? entry.execution.outcome,
+    expectsCompletionMessage:
+      entry.delivery?.payload?.expectsCompletionMessage ?? entry.expectsCompletionMessage,
+    spawnMode: entry.delivery?.payload?.spawnMode ?? entry.spawnMode,
+    wakeOnDescendantSettle:
+      entry.delivery?.payload?.wakeOnDescendantSettle ?? entry.wakeOnDescendantSettle,
+    terminalReply: entry.delivery?.payload?.terminalReply ?? entry.completion?.terminalReply,
   };
-}
+};
+
+export const markPendingFinalDelivery = (args: { entry: SubagentRunRecord; error?: string }) => {
+  const now = Date.now();
+  const payload: PendingFinalDeliveryPayload = loadPendingFinalDeliveryPayload(args.entry);
+
+  const delivery = ensureDeliveryState(args.entry);
+  delivery.status = "pending";
+  delivery.createdAt ??= now;
+  delivery.lastAttemptAt = now;
+  delivery.attemptCount = (delivery.attemptCount ?? 0) + 1;
+  delivery.lastError = args.error ?? null;
+  delivery.payload = payload;
+};
+
+export const refreshPendingFinalDeliveryPayload = (entry: SubagentRunRecord): boolean => {
+  const delivery = entry.delivery;
+  if (
+    !delivery?.payload ||
+    delivery.status === "delivered" ||
+    typeof delivery.announcedAt === "number"
+  ) {
+    return false;
+  }
+  delivery.payload = {
+    ...delivery.payload,
+    startedAt: entry.execution.startedAt,
+    endedAt: entry.execution.endedAt,
+    outcome: entry.execution.outcome,
+    terminalReply: entry.completion?.terminalReply,
+  };
+  return true;
+};
